Move *.py under hypnotoad2 subdirectory, organise into subdirectories
Enables proper module packaging.
