+++ hypnotoad2/geqdsk/_fileutils.py
