+++ hypnotoad2/test_suite/test_fileutils.py
@@ -1,5 +1,5 @@
 
-from .. import _fileutils
+from ..geqdsk import _fileutils
 from io import StringIO
