+++ hypnotoad2/utils/critical.py
