+++ hypnotoad2/utils/dct_interpolation.py
@@ -20,7 +20,7 @@
 from scipy.fftpack import dct
 import numpy
 
-from .mesh import MultiLocationArray
+from ..core.mesh import MultiLocationArray
 
 class DCT_2D:
     """
